bug fix - int nfolds

+++ src/utils.py
@@ -149,12 +149,12 @@ def load_df_pred(pred_dir, class_to_index):
 
 
 def train_test_valid_split_patients_stratified(df_full, test_size, valid_size, random_seed):
-    splitter = StratifiedGroupKFold(n_splits=1/test_size, shuffle=True, random_state=random_seed)
+    splitter = StratifiedGroupKFold(n_splits=int(1/test_size), shuffle=True, random_state=random_seed)
     split = splitter.split(X=df_full, y=df_full['int_dis_to_tum'], groups=df_full['patient_id'])
     train_inds, test_inds = next(split)
     df_train = df_full.iloc[train_inds].reset_index(drop=True)
     df_test = df_full.iloc[test_inds].reset_index(drop=True)
-    splitter = StratifiedGroupKFold(n_splits=1/valid_size, shuffle=True, random_state=random_seed)
+    splitter = StratifiedGroupKFold(n_splits=int(1/valid_size), shuffle=True, random_state=random_seed)
     split = splitter.split(X=df_train, y=df_train['int_dis_to_tum'], groups=df_train['patient_id'])
     train_inds, valid_inds = next(split)
     df_valid = df_train.iloc[valid_inds].reset_index(drop=True)
